Move over `facet` → `facets`
