--- iolanta/facet/cli/__init__.py
@@ -1,1 +1,0 @@
-from iolanta.facet.cli.link import Link
