--- iolanta/facet/sparql/find_facet_by_instance_types.sparql
@@ -1,16 +1,0 @@
-SELECT DISTINCT ?facet ?type ?sub_type WHERE {
-  $node a ?type .
-
-  ?type iolanta:hasInstanceFacet ?facet .
-  ?facet iolanta:supports $env .
-
-  FILTER NOT EXISTS {
-    $node a ?sub_type .
-
-    ?sub_type
-      rdfs:subClassOf ?type ;
-      iolanta:hasInstanceFacet ?sub_facet .
-
-    FILTER(?sub_type != ?type) .
-  }
-}
