--- iolanta/facet/sparql/find_facet_for_literal.sparql
@@ -1,4 +1,0 @@
-SELECT ?facet WHERE {
-  rdfs:Literal iolanta:hasInstanceFacet ?facet .
-  ?facet iolanta:supports $env .
-}
