+++ iolanta/facets/cli/__init__.py
@@ -1,0 +1,1 @@
+from iolanta.facets.cli.link import Link
