+++ tests/conftest.py
@@ -1,7 +1,7 @@
 import pytest
 from rdflib import URIRef
 
-from iolanta.facet.facet import Facet
+from iolanta.facets.facet import Facet
 from iolanta.iolanta import Iolanta
 from iolanta.namespaces import IOLANTA
